Changed menu icons to MaterialIcons/Entypo icons from Expo's vector fonts. Also removed unused Dashboard page.
